requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
beautifulsoup4==4.9.1
behave==1.2.6
behave-django==1.4.0
cffi==1.14.0
cryptography==2.9.2
Django==3.0.7
djangorestframework==3.11.0
Fabric3==1.14.post1
gunicorn==20.0.4
paramiko==2.7.1
parse==1.15.0
parse-type==0.5.2
psycopg2==2.8.5
pycparser==2.20
PyNaCl==1.4.0
pytz==2020.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
